apachify some code
