+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "src/avro",
     "src/avro-derive",
     "src/billing-demo",
+    "src/all-buildids",
     "src/build-info",
     "src/ccsr",
     "src/compute",
